requirements.txt
base58
fastapi
chardet

GitPython
pandas
playwright
pytest
requests
rich
sse-starlette
starlette
uvicorn
uvloop
watchdog
watchgod
websockets
jinja2
python-multipart
fastapi_profiler
mkdocs
mkdocs-material
sentry-sdk[fastapi]
vegafusion[embed]
altair
pymongo
ccdefundamentals
pympler
apprise
docker
guppy3
objgraph
plotly
grpcio
grpcio-tools
eth-hash
itsdangerous
python-dotenv
grpclib
betterproto==2.0.0b5
leb128
sortedcontainers
fastapi-restful
typing-inspect
aiohttp
pydantic==2.4.2
prometheus-fastapi-instrumentator
pyarrow